Version 1.5
Updates to the README file.

Move usage of f_setup_dev so we are only doing stty against the device once per run.

f_validate_dev usage has been moved to prevent unnecessary re-use, like f_setup_dev.

f_receive has been removed and merged with f_xmit. This was done to open the GRID_DEV for reading prior to writing. If we wrote "0a 00" to the device without opening it for reading, there was a strong possibility that the command would be corrupted before it reached the grid. This is probably because the fast-and-loose way I was accessing the device via bash.

Add/fix/improve various comments.

Add a trap to make sure we kill all background processes, such as hexdump.